Fix it so that pins show up on IE.
In pinry.js, reference is made to image.width and image.height
properties, neither of which exist. They are used to compute the
target height of the image, by computing the aspect ratio
image.width/image.height. Unfortunately, undefined/undefined yields
NaN, and so we end up writing "height='NaN'" into the HTML.

It looks like webkit (at least chrome) ignores it when the height
attribute on an <img /> element is set to NaN, but Internet Explorer
treats it as zero. That's why it didn't look broken on other browsers.